app/requirements.txt
flask
mysql-connector-python

Flask-Mail